1. 修改版本号未2.0 2. 修改文件上传方式为分片上传 3. 新增文件仓库，支持秒传

+++ mos-server/src/main/java/mt/spring/mos/server/entity/dto/SignDto.java
@@ -11,5 +11,5 @@ public class SignDto {
 	private Long openId;
 	private String bucketName;
 	private String resourceId;
-	private Long expireSeconds;
+	private Integer expireSeconds;
 }
